+++ CHANGELOG.md
@@ -15,6 +15,8 @@ contents to a new release section (e.g. _[0.x]_).
 
 - `ComputeDomainOfAlgebraicExpression` plan isn't shown when `ComputeDomainAndSimplifyAlgebraicExpression` plan can be
   shown (e.g., for `[3 * 2 / (x - 1)(x - 2)]`) (PLUT-708)
+- Fixed rational equations which can be reduced to a polynomial one just by simplifying and expanding,
+  e.g. `([1/x]+2)x=[2/x]x` (PLUT-703)
 
 ### Changed